rename __test__ directories to __tests__
