+++ src/lib/components/ContentfulRichText/ContentfulRichText.test.ts
@@ -3,7 +3,7 @@ import { render } from "@testing-library/svelte";
 import { describe, it, expect } from "vitest";
 import "$lib/__test__/extendExpect.ts";
 import ContentfulRichText from "./ContentfulRichText.svelte";
-import documentTestCases from "./__test__/documents";
+import documentTestCases from "./__tests__/documents";
 
 describe("ContentfulRichText", () => {
   Object.entries(documentTestCases).forEach(([label, { document, expectedHTML }]) => {
